update: final push

+++ app/src/main/java/com/example/mtt_rental/dtmodel/RoomTypeVM.kt
@@ -4,7 +4,7 @@ data class RoomTypeVM(
     val idRoomType:String,
     val price: Long,
     val area:Long,
-    val maxRenter:Int,
+    val maxRenter:Long,
     val description:String,
     val roomList:List<RoomVM>,
     val roomServiceList:List<RoomServiceVM> = emptyList()
