+++ app/src/main/java/com/example/mtt_rental/model/RoomType.kt
@@ -3,7 +3,7 @@ package com.example.mtt_rental.model
 data class RoomType(
     val idRoomType: String = "",
     val idApartment: String = "",
-    val maxRenter: Int = 0,
+    val maxRenter: Long = 0,
     val price: Long = 0,
     val area: Long = 0,
     val description: String = ""
